feat: add Annonce & Reservation APIs + fix 15 bugs (security, entities, architecture)
- Add @EnableMethodSecurity to activate @PreAuthorize role checks
- Make GET /api/annonces/** public (no auth required to browse listings)
- Add PATCH to CORS allowed methods (fixes frontend status update calls)
- Fix all 6 repository ID types (Long -> Integer) to match entity definitions
- Add @CreationTimestamp to Avis, Paiement, Message entities
- Prevent self-registration as ADMIN (whitelist: LOCATAIRE, HOTE only)
- Add ownership check on GET /api/reservations/{id} (prevent data leaks)
- Validate reservation status against whitelist (EN_ATTENTE/CONFIRMEE/REFUSEE/ANNULEE)
- Prevent host from booking their own listing
- Fix N+1 query on getAllAnnonces via JOIN FETCH
- Throw AccessDeniedException (403) instead of RuntimeException (400) for auth errors
- Move UserRepository out of ReservationController into service layer
- Update .gitignore with target/, IDE, and OS entries
- Add AnnonceController, AnnonceService, AnnonceRequest DTO
- Add ReservationController, ReservationService, ReservationDTO, CreateReservationRequest DTO

+++ backend/com.renthub/src/main/java/com/renthub/repository/AvisRepository.java
@@ -3,5 +3,5 @@ package com.renthub.repository;
 import com.renthub.entity.Avis;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AvisRepository extends JpaRepository<Avis, Long> {
+public interface AvisRepository extends JpaRepository<Avis, Integer> {
 }
