+++ backend/com.renthub/src/main/java/com/renthub/repository/PaiementRepository.java
@@ -3,5 +3,5 @@ package com.renthub.repository;
 import com.renthub.entity.Paiement;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PaiementRepository extends JpaRepository<Paiement, Long> {
+public interface PaiementRepository extends JpaRepository<Paiement, Integer> {
 }
